Revert "[CB-19712] Extend the response with more information"
This reverts commit e2e2c1bfa77791ee95a92e0a3dcd650a96d03707.
