+++ datalake-api/src/main/java/com/sequenceiq/sdx/api/model/SdxBackupStatusResponse.java
@@ -1,7 +1,5 @@
 package com.sequenceiq.sdx.api.model;
 
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
@@ -22,18 +20,13 @@ public class SdxBackupStatusResponse {
     @ApiModelProperty(ModelDescriptions.OPERATION_STATUS_REASON)
     private String reason;
 
-    @ApiModelProperty(ModelDescriptions.BACKUP_INCLUDED_DATA)
-    private List<String> includedData;
+    public SdxBackupStatusResponse() {
+    }
 
-    @ApiModelProperty(ModelDescriptions.BACKUP_TIMESTAMP)
-    private String timestamp;
-
-    public SdxBackupStatusResponse(String operationId, String status, String reason, List<String> includedData, String timestamp) {
+    public SdxBackupStatusResponse(String operationId, String status, String reason) {
         this.operationId = operationId;
         this.status = status;
         this.reason = reason;
-        this.includedData = includedData;
-        this.timestamp = timestamp;
     }
 
     public String getOperationId() {
@@ -50,14 +43,6 @@ public class SdxBackupStatusResponse {
 
     public void setOperationId(String operationId) {
         this.operationId = operationId;
-    }
-
-    public List<String> getIncludedData() {
-        return includedData;
-    }
-
-    public String getTimestamp() {
-        return timestamp;
     }
 
     @Override
