+++ datalake-dr-connector/src/main/java/com/sequenceiq/cloudbreak/datalakedr/DatalakeDrClient.java
@@ -2,7 +2,6 @@ package com.sequenceiq.cloudbreak.datalakedr;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -27,7 +26,6 @@ import com.google.common.base.Strings;
 import com.sequenceiq.cloudbreak.datalakedr.config.DatalakeDrConfig;
 import com.sequenceiq.cloudbreak.datalakedr.converter.GrpcStatusResponseToDatalakeBackupRestoreStatusResponseConverter;
 import com.sequenceiq.cloudbreak.datalakedr.model.DatalakeBackupStatusResponse;
-import com.sequenceiq.cloudbreak.datalakedr.model.DatalakeOperationStatus;
 import com.sequenceiq.cloudbreak.datalakedr.model.DatalakeRestoreStatusResponse;
 import com.sequenceiq.cloudbreak.grpc.ManagedChannelWrapper;
 import com.sequenceiq.cloudbreak.grpc.altus.AltusMetadataInterceptor;
@@ -214,9 +212,9 @@ public class DatalakeDrClient {
         );
     }
 
-    public DatalakeRestoreStatusResponse getRestoreStatusByRestoreId(String datalakeName, String restoreId, String actorCrn) {
+    public DatalakeBackupStatusResponse getRestoreStatusByRestoreId(String datalakeName, String restoreId, String actorCrn) {
         if (!datalakeDrConfig.isConfigured()) {
-            return missingConnectorResponseOnRestore();
+            return missingConnectorResponseOnBackup();
         }
 
         checkNotNull(datalakeName);
@@ -338,14 +336,15 @@ public class DatalakeDrClient {
 
     private DatalakeBackupStatusResponse missingConnectorResponseOnBackup() {
         return new DatalakeBackupStatusResponse(UUID.randomUUID().toString(),
-                DatalakeOperationStatus.State.FAILED,
-                List.of(), "", NO_CONNECTOR_ERROR
+                DatalakeBackupStatusResponse.State.FAILED,
+                Optional.of(NO_CONNECTOR_ERROR)
         );
     }
 
     private DatalakeRestoreStatusResponse missingConnectorResponseOnRestore() {
         return new DatalakeRestoreStatusResponse(UUID.randomUUID().toString(), UUID.randomUUID().toString(),
-                DatalakeOperationStatus.State.FAILED,
-                NO_CONNECTOR_ERROR);
+                DatalakeBackupStatusResponse.State.FAILED,
+                Optional.of(NO_CONNECTOR_ERROR)
+        );
     }
 }
