+++ datalake/src/main/java/com/sequenceiq/datalake/flow/dr/backup/DatalakeBackupActions.java
@@ -103,10 +103,10 @@ public class DatalakeBackupActions {
                 variables.put(BACKUP_ID, backupStatusResponse.getBackupId());
                 variables.put(OPERATION_ID, backupStatusResponse.getBackupId());
                 payload.getDrStatus().setOperationId(backupStatusResponse.getBackupId());
-                if (backupStatusResponse.getState().isFailed()) {
-                    sendEvent(context, DATALAKE_BACKUP_FAILED_EVENT.event(), payload);
-                } else {
+                if (!backupStatusResponse.isFailed()) {
                     sendEvent(context, DatalakeDatabaseBackupStartEvent.from(payload, backupStatusResponse.getBackupId()));
+                } else {
+                    sendEvent(context, DATALAKE_BACKUP_FAILED_EVENT.event(), payload);
                 }
             }
